restructered code and files frontend
