+++ frontend/src/tournament/local_tournament.js
@@ -1,4 +1,5 @@
 
+import { showMessage } from "../utils/utils.js"
 
 
 export async function startLocalTournament(playerCount) {
