+++ frontend/src/tournament/tournament_api.js
@@ -1,4 +1,4 @@
-import { fetchWithRefreshAuth } from './usermanagement.js';
+import { fetchWithRefreshAuth } from '../users_friends/usermanagement.js';
 
 // Tournament API functions
