+++ frontend/src/utils/utils.js
@@ -1,0 +1,32 @@
+
+
+export function showMessage(message, type = 'info', duration = 2500) {
+  const containerId = 'Message-container';
+  let container = document.getElementById(containerId);
+
+  if (!container) {
+    container = document.createElement('div');
+    container.id = containerId;
+    container.className = 'fixed top-4 left-1/2 -translate-x-1/2 z-50 flex flex-col gap-2 pointer-events-none';
+
+    document.body.appendChild(container);
+  }
+
+  const typeClass = {
+    success: 'bg-green-600 border-green-400',
+    error: 'bg-red-600 border-red-400',
+    info: 'bg-blue-600 border-blue-400'
+  };
+
+  const Message = document.createElement('div');
+  Message.className = `pointer-events-auto text-white px-4 py-2 rounded-md border shadow-lg ${typeClass[type] || typeClass.info}`;
+  Message.textContent = message;
+  container.appendChild(Message);
+
+  setTimeout(() => {
+    Message.remove();
+    if (container && container.childElementCount === 0) {
+      container.remove();
+    }
+  }, duration);
+}
